Revert "[LoRA] introduce `LoraBaseMixin` to promote reusability. (#8670)"
This reverts commit a2071a1837642b8e7360a5d0ddce6852b51e51a4.
